backend/requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
sqlmodel==0.0.27
psycopg2-binary==2.9.11
python-dotenv==1.1.0
redis==7.1.0
fastapi-cache2==0.2.2